refactor: reorganise mapping files
