+++ subgraph.yaml
@@ -28,7 +28,8 @@ dataSources:
           handler: handleTokenRestored
         - event: Transfer(indexed address,indexed address,indexed uint256)
           handler: handleTransfer
-      file: ./src/mapping.ts
+      file: ./src/main.ts
+
 templates:
   - kind: ethereum/contract
     name: TitleEscrowCloneable
@@ -39,7 +40,7 @@ templates:
       kind: ethereum/events
       apiVersion: 0.0.5
       language: wasm/assemblyscript
-      file: ./src/mapping.ts
+      file: ./src/escrow.ts
       entities:
         - TitleEscrowCloneable
         - TradeTrustERC721
